global: expose updateQueryState
- updateQueryState exposed withState HOC
- test for exposed props form withState
- documentation
- fix for config expecting axios property
- fix circular dependency src/lib/store.js -> src/lib/state/reducers/index.js -> src/lib/state/reducers/query.js -> src/lib/store.js
- changed initialResultsState.loading to true to avoid double children render in ResultsLoader closes #101
- closes #99
