+++ docs/docs/components/with_state.md
@@ -3,11 +3,17 @@ id: with-state
 title: withState
 ---
 
-`withState` is a HOC component that passes the redux state to an external component.
+`withState` is a high order component, which is used to expose the redux state and
+actions to external components.
 
-The component receives the up-to-date state every time something is changed.
+The component receives the up-to-date state every time something is changed and
+through its props it gains access to
 
-> Do **not** mutate the state inside your wrapped component!
+- `currentResultsState`
+- `currentQueryState`
+- `updateQueryState`
+
+> Do **not** mutate directly the state inside your wrapped component. Instead, use the function `updateQueryState` to pass your new query.
 
 ## Usage
