+++ package.json
@@ -16,7 +16,7 @@
   },
   "dependencies": {},
   "peerDependencies": {
-    "axios": "^0.19.0",
+    "axios": "^0.19.2",
     "lodash": "^4.17.15",
     "node-sass": "^4.12.0",
     "qs": "^6.8.0",
@@ -32,7 +32,7 @@
   "devDependencies": {
     "@babel/cli": "^7.5.5",
     "@svgr/rollup": "^4.3.2",
-    "axios": "^0.19.0",
+    "axios": "^0.19.2",
     "axios-mock-adapter": "^1.17.0",
     "coveralls": "^3.0.7",
     "enzyme": "^3.10.0",
