+++ src/lib/storeConfig.js
@@ -1,0 +1,20 @@
+/*
+ * This file is part of React-SearchKit.
+ * Copyright (C) 2018-2020 CERN.
+ *
+ * React-SearchKit is free software; you can redistribute it and/or modify it
+ * under the terms of the MIT License; see LICENSE file for more details.
+ */
+
+export const INITIAL_STORE_STATE = {
+    queryString: '',
+    suggestions: [],
+    sortBy: null,
+    sortOrder: null,
+    page: 1,
+    size: 10,
+    filters: [],
+    layout: null,
+  };
+
+export const STORE_KEYS = Object.keys(INITIAL_STORE_STATE);
